requirements.txt
torch==2.3.1
torchaudio==2.3.1
pyannote.audio==3.3.2
python-dotenv==1.0.1
speechbrain==1.0.2
audio-separator==0.24.1
onnxruntime==1.20.1
yt-dlp
groq==0.13.0
numba==0.60.0
ffmpeg-python==0.2.0
gruut==2.2.3
encodec==0.1.1
gruut_lang_en==2.0.0
gruut_lang_de==2.0.1
gruut_lang_fr==2.0.2
gruut_lang_es==2.0.1
umap-learn==0.5.1
anyascii==0.3.2
coqpit==0.0.17
einops==0.8.0
g2pkk==0.1.2
inflect==7.4.0
num2words==0.5.13
pysbd==0.3.4
trainer==0.0.36
Unidecode==1.3.8
pypinyin==0.53
dateparser==1.1.8
jsonlines==1.2.0
networkx==2.8.8
pynndescent==0.5.13
python_crfsuite==0.9.11
SudachiDict_core==20250515
SudachiPy==0.6.9
jamo==0.4.1
g2pkk==0.1.2
bangla==0.0.5
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
dateparser==1.1.8
gruut-ipa==0.13.0
python_crfsuite
pynndescent==0.5.13
tzlocal==5.3.1
hangul_romanize==0.1.0
ascii-magic==2.3.0
tensorflow==2.16.1
jieba==0.42.1
grpcio-status==1.33.2 
scikit-image==0.23.2
nltk==3.9.1
gradio==5.8.0
faster-whisper==1.1.1
transformers==4.46.3
librosa==0.10.2.post1
protobuf==4.25.8
ml-dtypes==0.3.2
onnx==1.19.0
pynvml==12.0.0
nvidia-nvml-dev-cu12==12.9.79
